separate infrastructure test functions
